spectrometer/requirements.txt
numpy>=1.20.0
opencv-python-headless>=4.5.0
matplotlib>=3.5.0
paho-mqtt>=1.6.0
